requirements.txt
colour-science==0.4.6
numpy==2.1.2
Pillow==10.4.0
requests==2.32.3
matplotlib==3.9.2
